Gp19-123: extract employee list to own componant
