+++ src/User/ProfilePage.jsx
@@ -3,23 +3,11 @@ import { StyledCardHeading, StyledCard } from "../StyledFormComponents";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 import { noManager } from "../Constants/Constants";
+import EmployeeList from "./EmployeeList";
 
 export const StyledLink = styled(Link)`
-  color: #7a517d;
-`;
-
-const StyledEmployeesList = styled.ul`
-  margin: 0;
-  padding: 0;
-  list-style-type: none;
-`;
-const StyledEmployeesListItem = styled.li`
-  padding: 5px;
-`;
-const StyledEmployeesLink = styled(Link)`
-  display: inline-block;
   text-decoration: none;
-  color: #322d38;
+  color: #7a517d;
 `;
 
 export const StyledEditLink = styled(Link)`
@@ -99,18 +87,7 @@ const ProfilePage = ({
             Employees:
           </StyledCardHeading>
           {employees.length > 0 && (
-            <StyledEmployeesList className="employeeList">
-              {employees.map(employee => (
-                <StyledEmployeesListItem key={employee.id}>
-                  <StyledEmployeesLink
-                    className="employeeLink"
-                    to={`/users/${employee.id}`}
-                  >
-                    {`${employee.firstName} ${employee.lastName}`}
-                  </StyledEmployeesLink>
-                </StyledEmployeesListItem>
-              ))}
-            </StyledEmployeesList>
+            <EmployeeList className="employeeList" employees={employees} />
           )}
         </StyledCard>
       )}
